"ads2_m8_a1_1: 6/6 testcases passed. Cannot run hidden testcases as Server is unreachablecheck_style score: 0/1 "

+++ M8/1. Shortest Paths/Solution.java
@@ -27,7 +27,7 @@ final class Solution {
 			String[] paths = sc.nextLine().split(" ");
 			int source = bst.get(paths[0]);
 			djk = new DijkstrasSP(ewg, source);
-			System.out.println(djk.distance(bst.get(paths[1])));
+			System.out.println((int) djk.distance(bst.get(paths[1])));
 		}
 	}
 }
